test: PartitionStream construction helper
Mocking out query responses requires constructing a PartitionResponse
containing the set of PartitionStream, itself a stream of RecordBatch.
This nested stream of structures is required to enable a pull-based /
streaming query response, but makes testing difficult because the types
are hard to initialise.

This commit adds a helper macro make_partition_stream! which when
combined with make_batch! to initialise the inner RecordBatch instances,
reduces the developer burden when writing test code that interacts with
query responses:

    let stream = make_partition_stream!(
        PartitionId::new(1) => [
            make_batch!(
                Int64Array("a" => vec![1, 2, 3, 4, 5]),
                Float32Array("b" => vec![4.1, 4.2, 4.3, 4.4, 5.0]),
            ),
            make_batch!(
                Int64Array("c" => vec![1, 2, 3, 4, 5]),
            ),
        ],
        PartitionId::new(2) => [
            make_batch!(
                Float32Array("d" => vec![1.1, 2.2, 3.3, 4.4, 5.5]),
            ),
        ],
    );

The above yields a PartitionStream containing two partitions, with their
respective RecordBatch instances.