fix: WAL/buffer commit cancellation safety
Ensure that a write that is added to the WAL is always attempted to be
applied to the BufferTree.

This covers off the case of a user submitting a write, waiting long
enough for it to be added to the WAL buffer, and then disconnecting
before it is added to the BufferTree (and before they get a response).

This is a minor issue on its own, but fixing it is necessary for correct
reference counting of WAL files:

    https://github.com/influxdata/influxdb_iox/issues/6566

This also documents a low-risk opportunity for the WAL contents &
BufferTree to diverge, potentially leading to a crash-loop at startup:

    https://github.com/influxdata/influxdb_iox/issues/7111

In practice a crash loop is unlikely, as it would require broken
invariants elsewhere (no schema validation being applied).

Closes https://github.com/influxdata/influxdb_iox/issues/6281.