refactor: owned persisted SequenceNumberSet copy
Allow an owned SequenceNumberSet to be destructured from an owned
CompletedPersist notification.

Additionally allow an owned SequenceNumberSet to be obtained from a
(potentially) shared Arc<CompletedPersist> in a memory efficient / no
cloning way. A CompletedPersist notification is typically wrapped in an
Arc, and at time of writing, typically referenced only once.