refactor: disallow empty UpstreamSnapshot
This simplifies the semantics of upstream RPC endpoint snapshots; an
empty snapshot will never be returned, and therefore if the caller has a
snapshot, it contains at least one endpoint to attempt an RPC call
against.

This moves the "no upstream" error handling from a write-time error, to
a pre-write error.