refactor(router): Move NamespaceSchema into cache on put
By moving the namespace schema into the Put cache method and returning
the new value wrapped in an Arc, it allows for the cache to merge the
new schema and the existing schema without calling clone() on either.

This has a side effect of allowing the metrics and stats capture
behaviour to be achieved without leaking into the traits definition.