feat(ingester): coalesce partition fetch requests
Reduce N concurrent partition fetch requests for the same partition into
a single catalog query. This prevents multiple queries executing when
all but one result is thrown away.

This removes a potential request amplification when the catalog is
unavailable, where a number of queries for the same partition execute,
see an error, and retry forever (with backoff) until the catalog
recovers, while more catalog queries are started.