fix(garbage_collector): delay initial s3 checker loop, fix dryrun
This PR makes 3 improvements.

* It adds the configured sleep interval at the start of the object store
  checker to avoid issues with making a remote list immediately at
startup. We see issues with the s3 api.
* the --dry-run flag was stopping deletes of objects from object store,
  but the retention flagger was still making updates to the catalog.
These writes to the catalog are surprising when the --dry-run flag is
provided. Now, with --dry-run the catalog is not updated. The logging
instead says how many records would be updated because of retention.
* It decreases logging in should_delete of the checker as it will be
  extremely noisey when reporting files it skips. An internal
environment has 3.8 million parquet files, most of which would be
skipped.

* related to #7363
* fixes influxdata/idpe#17451