fix(ingester): always complete partition fetches
In #7559 I introduced code to coalesce N queries for a given partition
down to a single query, with the other callers (async) blocking to
obtain the shared result.

This is usually fine, but it is hypothetically possible for all callers
to start the partition lookup and abort (cancelling their shared poll
future) before the partition resolves, leaving it half complete.

The problem is this half-complete future MAY be holding a connection
from the catalog connection pool (internally a semaphore). This means
this connection is "stuck", lent out to this future that isn't making
progress (and may never make progress!). Given sufficient occurrences,
it's possible this could lead to starving the system of catalog
connections.

This fix is simple - spawn the resolve future into the runtime so it
completes irrespective of any caller polling it (they only poll for the
result, not to drive it forward).

NOTE: this isn't practically possible, because writes that traverse the
WAL DmlSink currently become uncancellable - this may not always be the
case however, so this underlying foot gun should be patched over!