feat(router): Additively merge tables in NamespaceCache
This commit adds additive merge behaviour for tables missing
from the new NamespaceCache entry, as well as moving calculation
of change stat metrics down to the in-memory implementation.

The metrics no longer distinguish between insert and update
caches OPs as a result of the change to the `put_schema()` interface.