chore: renaming compaction job concurrency to df concurrency (#7686)
This PR renames the CLI option and variables throughout the compactor to be `df_concurrency` (and similar) instead of `job_concurrency` (and similar).

From the perspective of the compactor, the most noteworthy consequence of this limit is on concurrency within Data Fusion.  This is the limit on the number of DataFusion jobs the compactor may start concurrently.